Added player select. The idea is that each player will get a slice of the deck. Provided each player selects a different player number, they will never draw the same cards.
A downside to this implementation is that each player, when selecting the same player number, will also get the same subset every time they play.  Solutions include:
Ignorning splitting. Each character has the same decks, shuffled differently. Tolerate overlaps.
Have one player send a seed to the others. The seed is applied to shuffle the decks _before_ splitting. Splits will be non-overlapping, and unique to that key. As a bonus, the game will be deterministic/replayable when the same key is used.
Connect to a network. Players enter a game key, and are given an incrementing player number. From here, the game procedes the same as when players manually set player numbers.
Full network control: Network controls turns, scoring, etc. Possibly returns cards one at a time, to prevent cheating.

The advantage of the current approach is that it closely mimics a real card game. After download, no network connectivity is required, and dividing the decks can be simulated by manually entering 2 small numerical fields to evenly and pseudorandomly split the deck.